interface realizations added

+++ helpers/geometric_average_helper.py
@@ -4,7 +4,7 @@ from typing import List
 class GeometricHelper:
 
     @staticmethod
-    def get_geometric_average(items: List[int]) -> float:
+    def get_geometric_average(items: List[float]) -> float:
         """Returns geometric average of int items in list."""
         multiply = 1
         n = len(items)
